requirements.txt
streamlit
openai
llama-index
langchain
langchain_community
PyPDF2 
google-generativeai 
markdown